Optionally do not ignore GNOME_RR_ERROR_NO_MATCHING_CONFIG
Normally this is not an error; it means that there was no stored configuration that
matched the current set of outputs.  In that case, we ignore the error and return
'success'.

However, in the case where we are reconfiguring the outputs due to a RANDR hotplug event,
we *do* need to know if there was no matching configuration, so that we can fall back
to doing what 'xrandr --auto' does.

Signed-off-by: Federico Mena Quintero <federico@novell.com>