Disable smart configuration on hotplug, and leave just autoconfiguration
We don't want smart configuration to happen when the user hits the 'Detect displays' button
in gnome-display-properties --- the resulting XRRGetScreenResources() will cause the X server
to re-probe the outputs, and we'll get a RANDR event, which makes us reconfigure the outputs.

It is weird to get the outputs reconfigured (resolutions changed, etc.) when you just hit a
'Detect displays' button.  So for now, we'll just do what 'xrandr --auto' does, and leave
the more sophisticated version for later.

Signed-off-by: Federico Mena Quintero <federico@novell.com>