randr - don't take too long in the D-Bus method implementations
We used to pop up the confirmation dialog from within the D-Bus method implementations.  Since the
timeout for the confirmation dialog is longer than the D-Bus timeout, this could cause the user
to see an error similar to 'did not receive a reply'.

Now we do the confirmation dialog in an idle handler so that we don't block the D-Bus methods.

Signed-off-by: Federico Mena Quintero <federico@novell.com>