Low disk space warning bug-fixes
Cast buf.f_frsize and buf.f_bavail to gint64 before multiplying
them to calculate the absolute free disk space. This prevents an
overflow on 32-bit platforms.

If the device path for a GUnixMountEntry does not begin with /dev,
don't assume the mount is virtual and then ignore it. In some cases,
this can cause a real mount to be ignored (eg, the device path
can be rootfs for the root filesystem). Ignore a pre-determined
list of FS types instead, for filtering out virtual devices.

Also don't display a warning if there is greater than 2GB of free
space remaining. This fixes a regression from the old low disk
space warning. Fixes bug #591153.