Add GConf keys for to customize the XF86Display and XF86RotateWindows hotkeys
Normally having those keysyms hardcoded is fine if X is configured to
emit them when the user presses the corresponding keys on laptops.
However, for testing purposes on desktop PCs which don't have those
hotkeys, it is useful to configure different hotkeys temporarily.
The keys can of course be changed as well on badly-configured laptops,
but this should be better fixed at the kernel and X level if possible.

Signed-off-by: Federico Mena Quintero <federico@novell.com>