bnc#578296 - Always set the position of outputs, even if they are already turned on
Normally we would only position outputs that were turned off, but we
actually need to do this for all outputs that will remain on.  Otherwise
they could retain their old positions, which is wrong when in 'other' mode
we retain a single external output with a non-zero X offset.

Based on a patch by Takashi Iwai <tiwai@novell.com>

Signed-off-by: Federico Mena Quintero <federico@novell.com>