bgo#610482 - Adjust XF86Display timestamps if they are out of order with RANDR timestamps
Some laptops seem to send XF86Display keypress events when you unsuspend them,
probably as a way to 'make Windows work'.  For some reason these events
come with timestamps that are out of order with respect to the
RANDR timestamps.  So we'll adjust the keypress timestamps so that
the events can be processed.

Signed-off-by: Federico Mena Quintero <federico@novell.com>