Enable smart configuration on hotplug
This commit toggles the default on the
conditional introduced in
commit 900d3ed34cb691f056e82bf3cbe9c40ebb87c9e9

That commit was fixing a bug with the "Detect Displays"
button.  Unfortunately, the fix has the side effect of
breaking user expectations on hotplug.

I believe the original bug can be fixed in a different
way with some trickery in the gnome-rr code in gnome-desktop.

Regardless, the bug in "Detect Displays" seems like the
lesser of two evils.