Support multiple smartcard drivers at once
Previously we would look for the first available and
loaded smartcard driver and only listen for events for cards
of that type.  That's not great in the face of multiple readers.

Furthermore, it breaks the one reader case sometimes, as well,
if we end up selecting a software token (and thus ignore the
hardware token).