xrandr: Simplify the default XRandR behaviour
Instead of using 3 separate configuration items, only use one that
will default to not touching the carefully crafted setup from Xorg.

This makes it possible to default to using a "dock" setup (where the
internal screen would be turned off if an external one is present),
or defaulting to "clone".

https://bugzilla.gnome.org/show_bug.cgi?id=634092