automount: disable automounting while screen is locked
Based on an initial patch by Martin Pitt <martin.pitt@ubuntu.com>

On the recent Shmoocon there was a presentation by Jon Larimer demonstrating
how to abuse vulnerabilities and bugs, or even just creating socially or
security compromising thumbnails in mounting and thumbnailing, which happens on
automounting USB drives. This is a particular issue when this happens on a
locked box where the attacker doesn't otherwise have access to the user
account:

  http://www.net-security.org/secworld.php?id=10544

Disable automounting if the GNOME screen saver is currently locked.

https://bugzilla.gnome.org/show_bug.cgi?id=642020