orientation: Poll for value changes on accelorometer
We can't easily check for XI2 events happening solely on the
accelerometer, so instead, we'll poll the accelerometer device
a few times to look for value changes in its properties.

We currently look every 150 msecs, 5 times in a row, and only
when we receive events through udev that the device has changed.
Usually, we'll get the new state at the first attempt, which means
we can ask for XRandR to switch us within 200 msecs.