power: Do not sleep-on-idle by default
We do not actually read the boolean keys in gnome-settings-daemon due to
a mixup when the UI was re-designed. We can just change the default of
the timeout key (that we do read...) to zero to change "never" to be
the default.

Note: we're not actually going to remove the unused keys from the schema
like we did in master, otherwise upgrading the g-s-d package whilst
the control center is open is going to make it explode.

Resolves https://bugzilla.gnome.org/show_bug.cgi?id=660395