color: Don't assign the same device more than once at startup
In the color plugin at startup there's a lot going on async, such as
registering devices and monitoring profiles for changes. As it's all happening
at the "same time", and possibly out-of-order we can't order things linearly in
the startup phase.

This could mean we connect to a device and then straight away get the device
list (which also connects to the device) which means we're doing something twice
we only really need to do once.

Do solve this, use a simple hash table to keep track of what devices are being
assigned at startup. A device gets removed from the cache as soon as it's been
connected to, so we don't end up processing old data.

This speeds up the color plugin start by ~280ms if you have two display devices.