media-keys: Redraw volume OSD when not composited
gnome-settings-daemon draws a correct osd window when you hit a
media key, but does not redraw that window on volume changes afterwards,
unless a compositing manager is running (tested with xcompmgr).

It looks like the culprit is gsd_osd_window_update_and_hide, which
has a description of "Queues the @window for immediate drawing, and
queues a timer to hide the window." but only queues a redraw if the
window is composited.

Removing the check for compositedness makes it work without a
compositing manager.

https://bugzilla.gnome.org/show_bug.cgi?id=660990