Start of bgo#663126 - Disable a laptop's display when the lid is closed
This is the start of a set of fixes to handle laptops with docking stations.
The biggest problem right now is that with a docked laptop, closing the lid
doesn't disable the laptop's display.  You get xinerama-mode spanning the
(closed) LCD and the external display, which is useless.

This commit makes the xrandr plugin depend on upower-glib.  It makes the various
functions which generate RANDR configurations turn off the laptop's display if
the lid is closed.

In the next commits, we'll actually monitor the lid's state, and make other changes
to make switching more robust.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>