Save the generated RANDR configuration whenever possible
It used to be that we would only save the configuration if you clicked the 'Apply' button
in the control center's Displays panel (or capplet.  or whatever you want to call it
this week.).  This led to a situation where leaving a display configured 'just right'
by simply using the XF86Display hotkey, may later result in that configuration *not* being
restored, as it was never saved.  Now things should be more consistent, since the
configuration that you 'just leave there' without further changes will actually be
the preferred one the next time you use the same set of monitors.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>