If lid is closed, turn off laptop's display when applying a stored configuration
We copy the logic from gnome_rr_config_apply_from_filename_with_time() to load
the file by hand, but when we find a matching configuration in the file,
we check the laptop's lid to see if the laptop's display should be
turned off.

While this patch effectively removes the only use of
gnome_rr_config_apply_from_filename_with_time(), it does the same thing by hand.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>