Ensure the laptop is the primary display in auto-configuration and xinerama
Consider this situation: you close the laptop's lid, so only the external display
remains enabled.  Thus, that external display gets set as the primary output, as
it is the leftmost one.

Then you open the lid again.  The laptop's display gets turned on.  But since
ther already *is* a primary output (the external one), it didn't get re-set to
be the laptop - so you get a leftmost laptop, rightmost external, with the
primary on the external, which is not what you expect.

This commit makes make_xinerama_setup() force the laptop to be
the primary display.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>