Take lid state notifications from D-Bus instead of a signal
We'll have g-s-d's power plugin inform us of lid state changes, instead of connecting
directly to the UpClient ourselves.  This way we will not race with the power plugin
for the actions that need to be taken.  Also, the power plugin will then take the decision
of whether to suspend the machine.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>