Call the XRANDR plugin asynchronously
We can't make a sync call because

a) the XRANDR plugin runs in the same process, but in another GMainContext than
the one used for our sync call.  So, it can't accept connections while the
sync call's innermost poll() is running.

b) I guess this means DBus sucks for intra-process communication.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>