connect_after to the UpClient::changed from the power plugin
This lets the xrandr plugin, which is the other user of UpClient::changed, to run
its signal handler *first*.  By the time the handler in the power plugin gets called,
all the RANDR outputs will be configured correctly and the power plugin will
be able to make its decision about whether to suspend the machine or not.

Doing this with signals instead of an async DBus calls simplifies things,
and removes the scary/racy asynchronicity of suspending vs. reconfiguring outputs.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>