Refresh the GnomeRRScreen when the laptop's lid is opened or closed
One common scenario may be:  you plug or unplug an external monitor, and you close or open
the laptop's lid.  At *that* time, we can afford to probe the RANDR outputs (which is
an expensive operation).  This way we can deal properly with not having hotplug detection;
we can make use of the user's delay in reacting after opening/closing the lid to
get a better view of the state of the outputs.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>