color: Use a heuristic to set the screen atom when there is no primary device
In GNOME 2.32 we used a heuristic to work out what output profile should be set
screen-wide.

When converting the color process to g-s-d for GNOME 3, we switched to using the
primary display as the key to which monitor is the most important. This works
for most people, but some video drivers do not set a primary display when using
'xrandr --auto' and hence do not get a screen atom set.

In this case, fall back to the old heuristic when there are no primary monitors
displayed. This fixes applications like GIMP that expect a screen atom to be set
to properly color manage the display.

NOTE: We can't change GnomeRR to always return a primary monitor using the
same heuristic, otherwise it stops the screen control center capplet from
assigning the primary device correctly.