a11y-keyboard: only update GSettings on user input
X sends a message to us to tell us when the accessx settings are
changed.  They can change for two reasons: because a client (like
ourselves) has changed them or because the user did some keystroke (like
pressing shift 5 times to enable stickykeys).

We want to update GSettings in response to user requests, but prevent
spurious updates caused by the X server notifying us of the updates that
we just did for ourselves, so check that the update is actually caused
by user input.

https://bugzilla.gnome.org/show_bug.cgi?id=668213