power: Use the kernel-supplied backlight type rather than a whitelist
For a long time we have shipped a whitelist that mixes a list of devices we
should use, and the order we should attempt to do them. Since kernel 2.6.37 we
have a 'type' attribute which specifies if the device is a firmware device,
platform or raw device. According to mjg59, we should attempt to connect to
devices in this order only and nuke the whitelist.

This ensures the whitelist stays up to date as new backlight device names appear
upstream in the kernel.