keyboard: fix numlock_set_xkb_state
num lock state is stored in an enumeration called
GsdNumLockState.  Prior to commit
89b476ec6196aaad5c65804c8ceade3a92d33e29, the old_state
variable was being used as a boolean despite being
an enumeration. That commit fixed most of the code
to use the proper enumeration values instead of
TRUE / FALSE.

It missed numlock_set_xkb_state though.  This commit
fixes that function, too.