wacom: Make touch devices insensitve to KEY_IS_ABSOLUTE changes
Pen devices are typically used in absolute mode, while touch
devices are used in relative mode. However, the tracking mode
of *both* currently depend on the value of KEY_IS_ABSOLUTE.
If a preference is set for absolute or relative it will be
applied to both devices.

This patch has touch devices ignore changes to KEY_IS_ABSOLUTE.
Ideally it would be nice to have this key available independently
for both pen and touch devices at some point in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=670655