xsettings: remove global 'settings' list
This looks like an abortive attempt to merge the settings list for the
settings managers for all screens into one.  Unfortunately, the code
that actually calls into the settings managers to do updates still
updated each one individually and a bunch of code in the manager itself
was still referring to the per-instance variable (which was never
removed).

Switch back to each manager having its own copy.