daemon: Drop dbus activation capability
Years ago, GDM activated gnome-settings-daemon via dbus as
part of its hand-rolled greeter session. These days it uses
gnome-session instead of a hand-rolled session, and so
gnome-settings-daemon no longer needs to be activatable.

Furthermore, when gnome-settings-daemon is activated, the only
way it functions is if it's subsequently given an "Awake" call.
This limitation was added because it was getting unintentionally
activated in none-GNOME sessions (when those sessions were hosting
gnome apps, like rhythmbox), causing all sorts of havoc.

This doesn't-function-until-Awake-called limitation means that if
settings daemon ever crashes and gets unintentionally dbus activated
before gnome-session has a chance to respawn it, then the whole desktop
breaks (since whatever accidentally activates it won't call Awake).
This actually happens in practice because gnome-shell's power indicator
tries to track the power plugin's object properties.

This commit removes the ability for gnome-settings-daemon to be dbus
activated. It only causes problems, and no longer solves anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=673586