gnome-settings-daemon: fix applying settings to newly added touchpads
In GTK+-3.4, touchpads no longer are classified as GDK_SOURCE_MOUSE,
they are GDK_SOURCE_TOUCHPAD, so we need to add a check for that
in the handling for newly added devices, or touchpads won't get
configured properly when returning from suspend.

Required GTK+ version is bumped to 3.3.18 for the GDK_SOURCE_TOUCHPAD
enum value.

https://bugzilla.gnome.org/show_bug.cgi?id=673595