power: update fallback status icon on icon state change
When the icon was being changed from one icon to another, the fallback
icon was not being updated.

This meant that unplugging or re-plugging AC power on OLPC laptops
was not updating the icon (even though the tooltip updated just fine).

Add a missing call to update the GtkStatusIcon when the internal GIcon
changes.

Fixes http://dev.laptop.org/ticket/11878

https://bugzilla.gnome.org/show_bug.cgi?id=677313