keyboard: Rename apply_xkb_layout() to make its purpose clearer
This renames apply_xkb_layout() to apply_xkb_settings() to make it
clear that we also apply the options there. The options can't be
skipped if the setting is empty because then there would be no way for
users to get rid of them when disabling all of them in g-c-c.

https://bugzilla.gnome.org/show_bug.cgi?id=683878