power and media-keys: Use logind for suspending and rebooting the system
Use the new logind features to suspend and resume but making sure we opt out
of logind handling the sleep and power keys, and also inhibiting for lid close
auto-suspend if there is an external monitor connected.

Also use a delay inihibit for logind so that we can do actions on suspend like
blanking the screen using the screensaver and also poking the screensaver on
resume.

https://bugzilla.gnome.org/show_bug.cgi?id=680689