input: Clarify hotplug command exit value handling
The code and the comments were confused as to whether 0 or 1 should
be the magic value to mean "don't configure this device any more".

We are now using an exit value of "1" to mean "don't touch this device",
as it requires the user explicitely saying that they don't want the
device touched anymore. An empty script shouldn't cause all the
configuration to suddenly break...

https://bugzilla.gnome.org/show_bug.cgi?id=674251