keyboard: Add likely IBus engine for some locales if empty setting
If there are no input sources configured and we're running under one
of some specific locales where it's highly likely that the user will
need/want a specific IBus source to type we just add it to the list.

LC_CTYPE is used to determine the locale for this purpose since many
users in these locales set LC_MESSAGES to English.

https://bugzilla.gnome.org/show_bug.cgi?id=685457