wacom: touchstrip and touchring without a modeswitch
is something that can be found on older tablets
(e.g. Cintiq 12WX).

gnome-settings-daemon/gnome-control-center would add the
touchring/touchstrip once the corresponding modeswitch is
added, so without a modeswitch no touchring/touchstrip is
added even if the device features one.

This patch fixes the issue by adding the touchring/touchstrip
based on presence of such controls regardless of the presence
of a corresponding modeswitch.

https://bugzilla.gnome.org/show_bug.cgi?id=686022