main: Ensure session registration happens before other idles
We're deferring some slower operations to idles, in plugins,
to have it happen after session registration and not slow down
login - but I've observed that the ordering is unreliable.

Sometimes we end up doing the RegisterClient call after the
initial set of idles.

Ensure that the session registration happens before other idles by
making the idle higher priority than the plugins.

https://bugzilla.gnome.org/show_bug.cgi?id=686579