keyboard: Prevent potential infinite loop
XKB would notify us in the same way if the lockedMods
changed because of a programmatic, or a physical/human change.

This causes us changing the Num-Lock state generating another
event on top of the one we just processed, and might cause
infinite loops and 100% CPU usage.

Instead, we now only apply the settings:
- on startup
- when remember-num-lock is changed to true

https://bugzilla.gnome.org/show_bug.cgi?id=679151

Conflicts:
	plugins/keyboard/gsd-keyboard-manager.c