power: fix a warning at startup
idle_configure() will call idle_is_session_inhibited(), which will warn
if the session proxy is not available. This will always be the case when
idle_configure() is called at startup, since we create it
asynchronously, so we'll always warn.
Remove the warning by moving the idle_configure() call to the session
proxy ready callback.

https://bugzilla.gnome.org/show_bug.cgi?id=688253