keyboard: Trigger input source switching on key press for Caps Lock
We usually only trigger the switch on key release because we don't
want to prevent applications from getting e.g. Ctrl+Shift+E if the
shortcut is Ctrl+Shift.

But, for shortcuts including Caps Lock we can trigger the switch on
key press since applications don't usually include Caps Lock in their
own shortcuts and this allows us to:

1) be more responsive;

2) trigger the switch even if the user is typing quickly and thus we
   are getting other key events;

3) reduce the time window during which the lock mask is active in the
   server.

https://bugzilla.gnome.org/show_bug.cgi?id=688350