wacom: Avoid infinite recursion in gsettings callback
On a tablet PC, EDID matching will fail as the monitor is not
from Wacom and heuristic matching will be used instead in
find_output().

find_output() will then write the results in gsettings which
can cause an infinite recursion since find_output() is being
called (indirectly) from the gsettings callback in g-s-d.

To avoid this, remove the call to set_display_by_output() in
find_output().

https://bugzilla.gnome.org/show_bug.cgi?id=688811