screensaver-proxy: Add new screesaver proxy plugin
Proxy the org.freedesktop.ScreenSaver D-Bus API to gnome-session
as needed.

Note that, as we are the one talking to gnome-session, automatically
removing inhibition cookies when the caller goes away won't work.
We must watch incoming inhibition requests, and make sure to clean up
and remove lingering cookies when the caller goes away.

API:
https://projects.kde.org/projects/kde/kde-workspace/repository/revisions/master/entry/ksmserver/screenlocker/dbus/org.freedesktop.ScreenSaver.xm
Documentation:
https://projects.kde.org/projects/kde/kde-workspace/repository/revisions/master/entry/ksmserver/screenlocker/interface.h

https://bugzilla.gnome.org/show_bug.cgi?id=689225