wacom: multiple mode-switch buttons mode selection
The Cintiq 24HD has three modes per side, but also
3 different buttons to select modes. So you can go
from mode 1 to mode 3 without cycling through mode
2, you just hit the individual button.

Currently, when you hit any of the mode switch
buttons it just cycles through them like the other
tablets do.

To allow selecting a given mode, check if there are
multiple mode-switch buttons present on the device
and if so set the mode based on the index of that
button in the list of mode-switches.

https://bugzilla.gnome.org/show_bug.cgi?id=690157