power: Don't double/quadruple/... timeouts
This code in idle_adjust_timeout is meant to handle the very unlikely
situation that login takes longer than the blank timeout. But it can
have unintended side-effects where it causes timeouts to become much
longer than configured, just because an idle_configure call was triggered
when the system was already idle.

Remove both idle_adjust_timeout and idle_adjust_timeout_blank on the
basis that a system where login takes longer than 600 seconds (the
default blank timeout) is not suitable for running GNOME in the first
place.

https://bugzilla.gnome.org/show_bug.cgi?id=691002