power: Switch off backlight before suspend
When suspending, switch off the backlight and turn it back on when
resuming.

Note that we need to capture the current backlight before DPMS Off,
otherwise xbacklight will return zero.

Note that this only has an effect on machines where the lowest backlight
level turns the screen off, such as MacBooks.

https://bugzilla.gnome.org/show_bug.cgi?id=691702