color: Set the MAPPING_device_id on the auto-created EDID profile
This allows us to remove a lot of code responsible for watching for new profiles
to be added to the daemon and for adding them to devices ourselves.
You need to be running a colord daemon that contains the commit
842b9ebfd307d632aa1ff902940a4416c0200f0c for this to work.

If an auto-edid profile was previously created without the mapping metadata then
it is recreated at session start time.