power: Add ability to override XRandR detection
For our tests. The only thing we know how to do is check whether
there's an external screen or not, so that's what we allow to override
by creating a GSD_MOCK_EXTERNAL_MONITOR file in the current directory
with either '0' (no external screen) or '1' (with external screen) as
the content. Send the SIGUSR1 signal to gsd-test-power to generate
a fake XRandR change event, and get the file re-read.

echo 1 > GSD_MOCK_EXTERNAL_MONITOR
killall -USR2 gsd-test-power