xrandr: Don't operate on a NULL object
I see a critical out of gnome_rr_config_set_clone in my session log.
It turns out all uses of that function in the xrandr plugin are
happening after possibly nulling the GnomeRRConfig they are operating
on.

This commit ensures the clone state is set immediately after creating
the object to correct that.

https://bugzilla.gnome.org/show_bug.cgi?id=692372