a11y-keyboard: Disable everything on exit if no settings changed
The default setting in GDM is to have the keys enablable
through keyboard shortcuts, and this gets inherited by the
user sessions.

The problem is that non-GNOME users won't have anything
changing those defaults back to their expected values
(so that Slow Keys don't triggered for example), and this
causes problems:
https://bugzilla.redhat.com/show_bug.cgi?id=816764

If the user doesn't change anything to the defaults in the GDM
session, we revert to the pre-"GNOME session" values.

https://bugzilla.gnome.org/show_bug.cgi?id=685063