updates: Fix crasher when firmware updates is disabled
g_ptr_array_remove_index_fast() was called as well as a free
function, but the GPtrArray already had its own free function.

Furthermore, g_ptr_array_remove_index_fast() replaces the
indexed element with the last element, which means the current index
must still be checked, but the for loop skipped it.

https://bugzilla.gnome.org/show_bug.cgi?id=694129