Revert "cursor: Only enable the OSK when touch is used"
This reverts commit 2c5e3f8ac3803a19de49d09d1c028b75f7b048d2.

The code as it is is broken, as we don't differentiate between
"no cursor because we haven't used the device yet" and "a touchscreen
has been used".

It also changes a GSetting which is something we should not be touching
(gnome-shell or gnome-settings-daemon should be using the GSetting in
addition to this computed state to show the keyboard or not).