power: Don't change active state variable in VM
The intent of the virtual machine detection was to avoid changing
power states; however, while we avoided some parts of the code, we
still ended up flipping the state variable to which other parts of the
code reacted.

With this patch, in a VM, we remain GSD_POWER_IDLE_MODE_NORMAL
permanently.

https://bugzilla.gnome.org/show_bug.cgi?id=696500