media-keys: fix race condition
For some reason I see a crash in the settings daemon in the
gsettings_changed_cb() callback when accessing manager->priv->keys->len.

It seems that the gsettings callback is called before
manager->priv->keys is initialized. And that would happen after
because init_kbd() which initialized manager->priv->keys is called
upon shell's dbus appearance.

https://bugzilla.gnome.org/show_bug.cgi?id=697698