power: Don't poll for external monitors
Instead of checking every 30 seconds whether we should suspend,
only start the timeout if we should suspend but don't because
the external monitor is gone.

We currently ignore changes to "default-monitors-setup" because it's
unlikely that they will happen between external monitor changes.

This fix also ensures that there are really 30 seconds of safety timer
instead of somewhere between 0 and 30 seconds.

https://bugzilla.gnome.org/show_bug.cgi?id=698132