power: Remove pre-dim brightness hack
We were trying to work-around the fact that switching to the
VT on some laptops re-enabled the backlight.

The backlight level is however incorrect when DPMS mode off, and
checking the level of the backlight, when DPMS is still on, caused
races.

Linux kernel 3.10 will remove the need for VT switching which will
fix this bug properly:
http://blog.ffwll.ch/2013/04/neat-drmi915-stuff-for-310.html

https://bugzilla.gnome.org/show_bug.cgi?id=696379

Conflicts:
	plugins/power/test.py