power: Use a DBus property for screen brightness
A property is easier for clients to manage, as they'll automatically
get new PropertiesChanged signals when the value changes, and with
GDBus bindings, properties are cached field lookups that don't require
asynchronous access.

If there are any errors when fetching the new brightness, the value
-1 is returned, so the interface has been modified slightly to support
signed values.

Also adapt the color and media-keys plugins to the new interface.

https://bugzilla.gnome.org/show_bug.cgi?id=698754