color: Do not try to parse .goutputstream files
When cd_icc_save_file() is called it uses g_file_replace_contents().
This creates the .goutputstream momentarily which is picked up by the directory
watch on ~/.local/share/icc.

Getting the info never succeeds as the files have vanished before they can be
read and a nasty error is stored in the journal.