wacom: Add gsd_wacom_oled_gdkpixbuf_to_base64 function
Add function that encodes GdkPixbuf to base64 string. The returned
string starts with MAGIC_BASE64 "base64:". The pixbuf has to be
64x32 pixels otherwise function returns NULL. The function averages
pixel value, so for grayscale GdkPixbuf there is no change and it
should make color icons to look acceptable on wacom OLED screens.

The function is not (yet) used.

https://bugzilla.gnome.org/show_bug.cgi?id=700499