power: Make GetPrimaryDevice always return the composite device
With one battery, the composite device should report what we want as
well, so this shouldn't see any major changes in most cases. The reason
this is wanted is so that when the device is charging or fully charged,
we'll get the composite device as well, rather than throwing an error.

https://bugzilla.gnome.org/show_bug.cgi?id=700913