xsettings: Add support for window-scaling
We read the scale-factor gsettings and propagate to gdk, including doing
auto-detection if the setting is 0.

NOTE: The auto-detection code is currently disabled as enabling
scaling breaks gnome-shell. When this is fixed we want to enable this.

We also scale the old Xft.DPI setting by the window scale so that old
applications get a decent size on very high dpi monitors, plus we report the
unscaled dpi so that window scale aware apps (like gtk) can avoid this
dpi scaling.

We also scale the cursor theme size so that you will get matching sizes
for cursors.