xsettings: Add support for window-scaling
We read the window-scale-factor gsettings and propagate to gdk, including doing
auto-detection if the setting is 0.

Additionally, we scale the old Xft.DPI setting by the window scale so that old
applications get a decent size on very high dpi monitors, plus we report the
unscaled dpi so that window scale aware apps (like gtk) can avoid this
dpi scaling.