xrandr: don't ever show a dialog for xrandr failures
First, we're using gtk_dialog_run which is a big no-no in a system
service (since it blocks the service).

Second, a user will be able to tell if their configuration changes
aren't working, the dialog doesn't help them.

Third, the dialogs frequently contain technobable about crtc 65
and the like.

This commit gets rid of them.

https://bugzilla.gnome.org/show_bug.cgi?id=704473