media-keys: Use MPRIS when no apps use the media-keys API
If the media-keys plugin gets a Play or Pause key events and we
don't have any clients registered to handle it through the native
media-keys API, look for an MPRIS interface using bus namespace
watching and try that.

This fixes integration of Spotify for example, as it's unlikely to
implement a GNOME specific interface.

https://bugzilla.gnome.org/show_bug.cgi?id=697810