common: rename gnome-settings-session to gnome-settings-bus
The gnome-settings-session code is so named because it used to
exclusively house login session specific code (ConsoleKit/logind/gnome-session).

Since then it's been used more generally as a place to stuff singleton
dbus proxies used across multiple plugins.

This commit renames it to gnome-settings-bus to give it a more
appropriate name for its current role.

https://bugzilla.gnome.org/show_bug.cgi?id=705127