color: Rename GsdColorX11 to GsdColorState as it's no longer only for Xorg
Weston embeds all the colord functionality into the compositor itself.
For GNOME Shell, Giovanni has decided to abstract out the GnomeRR calls to call
into a new private DBus interface. This means we have to continue to use the
GnomeRR calls in g-s-d and just avoid doing the X11 functions.