xrandr: adapt to the new model for display configuration
Stop handling monitors.xml and the initial modeset for the session,
as well as monitor hotplug and lid switches. All that is implemented
in mutter now.
What is left in the xrandr plugin is Fn-F7 (for lack of a better
place) and Rotate. Both will go in mutter/gnome-shell at some point.

https://bugzilla.gnome.org/show_bug.cgi?id=705507