updates: Notify about available offline updates
gnome-shell no longer shows a menuitem when a prepared offline
update is available, so the updates plugin needs to take over
the responsibility and pop up a notification in this case.

We monitor /var/lib/PackageKit/prepared-update and show a notification
when the file is created. We don't show it more often than once
every 5 minutes. To handle the case that the file already exists
when gnome-settings-daemon is started, we check for its existence
once, 5 minutes after starting. The notification offers to launch
gnome-software to view details of the prepared update.

If gnome-software is not available, we just trigger the restart
to install the updates directly from the notification.

https://bugzilla.gnome.org/show_bug.cgi?id=706696