datetime: Add new datetime plugin
For now, this takes care of automatic timezone updating based on
geolocation. It relies on the new GeoClue2 D-Bus service for getting the
location information and updates the system TZ using the
org.freedesktop.timedate1 service.

In the future, it is also going to take care of displaying notifications
for upcoming time related changes ("Daylight Saving Starts Tonight") and
notifying the user about timezone changes.

https://bugzilla.gnome.org/show_bug.cgi?id=706979