screensaver-proxy: Fix crash when client vanishes
We were using the "sender" variable, which was gathered from the
hash table, instead of using directly the argument passed to the
function. As the sender variable was freed pretty much straight
after it was removed from the cookies hashtable, we were accessing
freed memory, usually leading to a crash.

Reported in:
https://trac.videolan.org/vlc/ticket/4739#comment:46