build: Depend on geocode-glib 3.10.0
This makes sure to avoid confusion which geocode-glib release we really
need as the pre-release versions used different numbering.

Also, we need latest 3.10.0 to make sure to pull in json-glib that's
built with the -Bsymbolic linker flag. This ensures that g-s-d doesn't
crash when both json-c and json-glib are in the same process.

http://blogs.gnome.org/ebassi/2013/09/16/do-not-link-against-pulseaudio-and-json-glib-0-16/