power: Fix time to empty with multiple batteries
If one battery is draining and the other one isn't, the time
to empty wouldn't be zero, but it would only match the time
to empty for the single battery.

Instead, short-circuit the time_to_empty re-calculation for
single battery systems, and ignore the accumulated time to
empty/time to full for multiple batteries and recalculate it.

https://bugzilla.gnome.org/show_bug.cgi?id=710344