xsettings: export Gtk/ShellShowsDesktop setting
Export a property for whether the desktop icons are shown or not.  This
will allow Gtk to decide if we should show the 'Desktop' item in the
places sidebar, according to if the user has enabled their desktop icons
or not.  We use the value of the GSettings key 'show-desktop-icons' from
'org.gnome.desktop.background'.

https://bugzilla.gnome.org/show_bug.cgi?id=712302