wacom: Add work-around for configs on NFS shared homes
When sharing the same home directory over NFS, it might be convenient
to share device configurations as well, so that the "Wacom Intuos X" on
machine A has the same configuration as "Wacom Intuos X" on machine B
for that user.

The work-around makes the machine-id be all zeroes when
~/.config/gnome-settings-daemon/no-per-machine-config
and as such, should not be applied during a running GNOME session.

https://bugzilla.gnome.org/show_bug.cgi?id=723099