keyboard: Apply num-lock to newly connected keyboards
before 34395459cc8f0da6f163736743bced441ef86390, apply_all_settings()
and then apply_numlock() would sync up the numlock LED on a newly
hotplugged device.
34395459cc8f0da6f163736743bced441ef86390 removed that call, so devices
now come up with the numlock LED off even when the numlock is on, and a
subsequent state change will wrongly toggle it, i.e. LED off when numlock
is on, LED on when numlock is off.

This should really be fixed in the xserver but it's unlikely to happen,
restoring this patch in gnome-settings-daemon seems the simplest solution
for now.

https://bugzilla.gnome.org/show_bug.cgi?id=722753