main: don't die when gnome-session says Stop
gnome-session asks all registered clients to stop after the user
confirms the poweroff/logout dialog, but we should ignore that
request, because non registered applications are still mapped
and they would lose their theme and icons if we die (and same
for the shell).
We will go away as soon as the X11 connection is closed or the
session bus dies anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=707790