xrandr: Use GsdDeviceMapper to rotate touscreens
The Xrandr module was previously in charge of rotating touchscreens,
although it did so in a limited number of cases (eg. only through fn-F7,
not through the g-c-c panel), and didn't do anything to map touchscreens
to the right output if multiple monitors are present, so touchscreens
would be left mapped to the whole span of monitors.

Fix this by using GsdDeviceMapper, so touchscreen mapping is managed by
that object. The Xrandr module still gets to manage what/whether
touchscreens get added to the mapper.

https://bugzilla.gnome.org/show_bug.cgi?id=709600