GsdDeviceMapper: use an async constructor for GnomeRRScreen
GnomeRRScreen depends on mutter, so it cannot be created synchronously,
because doing so blocks gnome-settings-daemon from registering with
gnome-session, and thus blocks mutter/gnome-shell from being
spawned. Which triggers a timeout and fails.
Instead, do the init asynchronously. We lose the error reporting,
but in practice the call should never fail. Also, this creates
a small window of time in which GsdDeviceMapper sees no outputs,
but that's ok, because we recompute everything when the screen
is ready.

https://bugzilla.gnome.org/show_bug.cgi?id=724323