wacom: add Bluetooth OLED handling for Intuos4 WL
That patch implements OLED handling over bluetooth for Intuos4 Wireless
tablets. Due to the nature of the device which accepts 4 bit colour images over
USB and only 1 bit colour images we have to handle those 2 situations
differently.

On top of that kernel side has been implemented in such way that
the images for USB connection have to be scrambled in userland before they are
send to the device.

The decision to move scrambling to the OLED helper and to
have 4-bit to 1-bit conversion in the same location is the cleanest solution
that I found to the problem of inconsistent image formatting. It changes
the previous approach to keep scrambling out of the helper which is executed
with elevated privileges.

https://bugzilla.gnome.org/show_bug.cgi?id=724955