xsettings: Don't mix GnomeRR and GDK information
If we don't mix information from GnomeRR and GDK, then the "changed"
signal on GnomeRRScreen is all we need to properly know when the
resolution of the system is changed - mixing them presents problems
because GDK may not yet have updated it's information from X when the
"changed" is triggered by a D-Bus signal.

We still use GDK information before we have a GnomeRRScreen to minimize
the chance of switching the window scale during startup.

https://bugzilla.gnome.org/show_bug.cgi?id=709859