datetime: Call geoclue's stop method when we no longer need it
This makes sure the geolocation marker in gnome-shell's status area goes
away after turning Automatic Time Zone off.

Normally, geoclue would watch for the clients to disappear on the bus,
but since we are using a shared singleton connection to the system bus
in gnome-settings-daemon, other plugins can still keep the connection
alive after we unref the dbus connection in the datetime plugin. To let
geoclue know we no longer need it, explicitly call the DBus Stop method
when stopping geolocation.

https://bugzilla.gnome.org/show_bug.cgi?id=723932