power: Check that the output is connected when determining if it's ON
It might happen that there's no device connected and yet the CRTC
associated with an output has a valid mode.

RRGetOutputInfo tells us if the hardware was able to detect a device
connected to the output so let's use that.

Note that in versions newer than 3.8, this check isn't required since
the list of outputs that we get from mutter's DisplayConfig interface
is already filtered by mutter and doesn't contain any disconnected
outputs.

https://bugzilla.redhat.com/show_bug.cgi?id=1081093