power: Fix wakeup with some keymaps
When trying to wake up the screen, try to use XF86WakeUp in preference
to the left Alt. Don't use the right Alt key as it might be set as the
Compose key which would eat the event without resetting the idle.

Also make sure to initialise the keycode only once.

https://bugzilla.gnome.org/show_bug.cgi?id=729375