a11y-keyboard: Make sticky and slow keys notifications critical
Since these features change the way the keyboard behaves we shouldn't
timeout the notifications automatically and leave users confused,
thinking their keyboard is broken, in case they activate them
unknowingly and miss the notification.

We use the critical urgency level to achieve this since it makes
gnome-shell automatically expand the notification, allowing users to
act on the actions immediately, and it also means that the
notification won't disappear until it's explicitly acknowledged.

https://bugzilla.gnome.org/show_bug.cgi?id=729714