wacom: Make OSD always appear on a single monitor
If the tablet has no monitor assigned, the button mapping OSD would cover
the whole span of monitors, which looks somewhat broken on multimonitor
(edit button in the middle of virtual space, svg image spread across both
monitors, etc...).

Change this behavior to always pick the primary monitor if no monitor is
assigned to the device, this makes the OSD look less broken.

https://bugzilla.gnome.org/show_bug.cgi?id=732110