daemon: Fix crash when a plugin failed to start
When a plugin failed to start, and wouldn't return an error, we'd
crash trying to access the error. The error wasn't immediately
visible as this part of the code is implemented as one large macro.

This fixes a crash on startup when gnome-settings-daemon is compiled
with systemd support but logind isn't available on the running system.

https://bugzilla.gnome.org/show_bug.cgi?id=732836