orientation: Fix plugin on MS Surface devices
When plugging in the type/touch cover on a Microsoft Surface, the
IIO sensor is unplugged, and plugged in again with a different
sysfs path. As the accelerometer comes and goes, we need to handle this
in the orientation plugin, and not bail out if we can't find the
accelerometer on startup, and handle devices coming and going.

https://bugzilla.gnome.org/show_bug.cgi?id=733507