media-keys: Postpone initialisation of libcanberra
It's possible for gnome-settings-daemon to block for the initialisation
of PulseAudio on startup, especially as some of PulseAudio's
enumerations could take a while.

Instead of initialising libcanberra on startup, initialise it when we're
actually going to use it the first time. Given that this happens in a
volume media key handling function, there's a good chance that
PulseAudio will already be initialised at that point.

See https://bugzilla.gnome.org/show_bug.cgi?id=645756#c3

https://bugzilla.gnome.org/show_bug.cgi?id=735777