xsettings: Fix swapping out of WM settings schemas
Until we get session-specific defaults in GSettings, we pick up the
button-layout setting from different schemas depending on whether we
are running in classic mode or not. However there is a serious thinko
in the current implementation: When running in classic mode, the
settings value is swapped with the one from the classic schema,
otherwise it is used directly - which turns out to be wrong when
the value itself is taken from the classic schema, e.g. when receiving a
change notification (which apparently may even be the case randomly
at startup).
Avoid this confusion by not trying to add the classic schema unless
actually running in a classic session.

https://bugzilla.gnome.org/show_bug.cgi?id=736974