daemon: Add a check for wayland sessions
We want to do some things differently, or disable some features
entirely, when running under a wayland session.

Adding a global function which tries to connect to a wayland
compositor using the same method as a regular wayland client allows us
to achieve that.

For now, since we don't need to do anything else with the wayland
connection we tear it down immediately after checking. Note that g-s-d
uses the X11 GDK backend exclusively so there's no danger of this
being a duplicated wayland connection.

This commit introduces an optional build time dependency on
libwayland-client which is, by default, enabled or disabled
automatically depending on the pc file existence.

https://bugzilla.gnome.org/show_bug.cgi?id=738009