media-keys: Disallow screenshots when locked
Allowing random people to create files in the user's home folder
while the screen is locked is unexpected at best, so block the
corresponding shortcuts (while still allowing taking screenshots
to the clipboard for bug reports/reviews etc.).
While at it, adjust the shortcut for screencasts as well (though
the change their is purely informational, as screencasts are already
blocked by the shell while locked).

https://bugzilla.gnome.org/show_bug.cgi?id=737456