common: Add GsdDeviceManager
This is a GUdev-based object that keeps track of the plugged in input
devices. Meant to be used across g-s-d wherever some knowledge about these
must remain.

On X11, some API is still offered to access the gdk/X devices (which may not
map 1:1 with the udev devices).

For both platforms, the relevant /org/gnome/desktop/peripherals/... settings
can be accessed, in order to interoperate with the settings that mutter
applies on libinput devices.

https://bugzilla.gnome.org/show_bug.cgi?id=742593