common: Make GsdDeviceMapper use devices from GsdDeviceManager
This has brought some changes in the xrandr and wacom modules, which now
use GsdDevice for device mapping.

The use of GdkDevices and XInput properties remain for these settings not
yet in the gsettings-desktop-schemas peripherals settings in the wacom case,
and those bits have been made to run only on X11 as expected.

https://bugzilla.gnome.org/show_bug.cgi?id=742593