mouse: Use gsettings-desktop-schemas mouse/touchpad settings
Most settings have been moved there, so make this plugin use the new
schemas for the handled settings, and remain using the settings-daemon
schema for settings that can just be honored on X11.

In order to simplify the situation for the future, the management of
"middle-button-enabled", "horiz-scroll-enabled" and "disable-while-typing"
settings have been removed, and the sane default is applied on non-libinput
devices on X11.

For all other settings, the mouse plugin has been made to ignore devices
created through the libinput driver, as mutter handles these.

Additionally, the entire plugin is disabled on Wayland, the settings
remaining in g-s-d schemas are not a thing there.

https://bugzilla.gnome.org/show_bug.cgi?id=742593