common: Migrate input settings to gsettings-desktop-schemas keys
The deprecated keys are kept in isolated ".deprecated" suffixed schemas.
On plugin startup, the settings-daemon paths are opened with these legacy
schemas, and all user-modified keys are dumped to the new location and
reset.

This ensures the transition just happens once per-key, and ensures the
dconf database is left clean.

https://bugzilla.gnome.org/show_bug.cgi?id=742593