power: On raw backlight types, clamp the value to a minumum
On raw backlight types, it's likely that setting it to zero will
completely turn the backlight off which isn't very useful. Instead,
let's always clamp the value to a minimum in that case.

For now we just clamp if the backlight has at least 100 levels but we
might tweak it as needed.

https://bugzilla.gnome.org/show_bug.cgi?id=744278