power: Save last requested idle mode while temporarily unidle
If one of our idle watches triggers while we're temporarily unidle we
should save and switch to it when that period ends otherwise we would
ignore any idle request transitions in that period and always go back
to the state where we were when becoming temporarily unidle.

This prevents the following bug:

1. Laptop enters sleep
2. Lid open, wake up, idle time gets reset, we enter NORMAL state
3. There's a notification so gnome-shell asks us to go temporarily
   unidle while we're in NORMAL state
4. The 15s inactivity watch is triggered and we switch to BLANK
5. The unidle timer is triggered and we switch back to NORMAL since
   that's the state we were in when it started

Result is that the screen is left turned ON indefinitely after waking
from sleep if there's a notification when we wake up.

https://bugzilla.gnome.org/show_bug.cgi?id=748849