device-mapper: Reset the "display" setting on NULL outputs
We used to attempt to reconstruct the ['','',''] default, although
this failed on NULL outputs as we tried to feed NULLs into "as"
variants/settings.

Fix this by just resetting the key when the device gets a NULL
output, we can also make the other path clearer as we can't get
NULLs in edid[] as we previously might.

Based on a patch by Alexandre Rostovtsev <tetromino@gentoo.org>

https://bugzilla.gnome.org/show_bug.cgi?id=749844