smartcard: use NSS_InitContext instead of NSS_Initialize
NSS_Initialize is a noop if called multiple times.  We
currently call NSS_Initialize twice in gnome-settings-daemon.
Once by NMClient and once by the smartcard plugin.  NMClient
does it first, and it does it without initializing the secmod
database. When the smartcard plugin tries to initialize NSS
with the secmod database later, it's call is turned to a noop.

This commit changes the smartcard plugin to use NSS_InitContext
instead, which can properly handle being initialized multiple
times with different configurations.  See:

https://wiki.mozilla.org/NSS_Library_Init

https://bugzilla.gnome.org/show_bug.cgi?id=751040