media-keys: Remove "button-*" configuration
The button-power, button-suspend, etc. configurations accept a
GsdPowerActionType, but not all of the variants are supported:
        case GSD_POWER_ACTION_BLANK:
        case GSD_POWER_ACTION_LOGOUT:
        case GSD_POWER_ACTION_NOTHING:
                /* these actions cannot be handled by media-keys and
                 * are not used in this context */
                break;

Given that we usually prefer to avoid this sort of configuration which should
work out of the box, it would be better to have them act as they are labelled,
and people who want to change the configurations at system level can use udev to
apply a new keymap.

https://bugzilla.gnome.org/show_bug.cgi?id=753713