wacom: Apply left handed mode
With the introduction of GsDeviceMapper (commit 0810de0f), left-handed
mode became responsibility of this object, which was applied by applying an
extra rotation to the device coordinate matrix.

Later, the device coordinate matrix handling was lifted to mutter, although
the rotation setting was kept in the g-s-d side of the settings split, so
GsdDeviceMapper would listen to the wrong settings, and left-handed mode
became no mans land.

Put the application of this setting back into GsdWacomManager, again
translated to the "Wacom Rotation" device property. Nonetheless, all
settings should move eventually to mutter domain, and this turned again
into modifications to the coordinate matrix.

https://bugzilla.gnome.org/show_bug.cgi?id=749767