common: Add GsdDeviceManager::device-changed
When devices are plugged, we may end up emitting
GsdDeviceManager::device-added at a time when not all GdkDevices have been
attached to the GsdDevice, this makes these devices occasionally not being
applied the right keep-aspect/area settings.

Add a GsdDeviceManager::device-changed signal, emitted for the GsdDevice
for those situations, so the late-added devices are ensured to be updated.

https://bugzilla.gnome.org/show_bug.cgi?id=753589