media-keys: Re-add configuration for the power button
In commit 50564cde49ca2e17fb7e59f36a35d61c2cbef1af, we removed support
for configuring the various "sleep state" buttons. However, the power
button might need different behaviour based on the machine, which we
cannot always detect.

In later commits, we'll hard-code the actions for tablets and virtual
machines. Note that "power off" is not an option as this would make
the default action too destructive. It is recommended that you use a
custom shortcut for this instead.

https://bugzilla.gnome.org/show_bug.cgi?id=755953