xrandr: Avoid crash when Fn+F7 is pressed on some systems
When a display driver doesn't support XRandR properly,
manager->priv->rw_screen will be NULL. However, because of the way that
we need to initialise D-Bus, some of the API is still accessible.

As such, the "Fn+F7" (update screen) shortcut would access API that
should only be accessible when XRandR is available. Add guards to those
functions.

https://bugzilla.redhat.com/show_bug.cgi?id=1196044