media-keys: Use new gvc and g-s APIs for audio device selection
When plugging in a headset, or a microphone in some computers, they
will not be able to detect that there is a new input available (for
the headset), or that it doesn't have any outputs (for the
microphone).

Using gvc's new audio-device-selection-needed signal we call into
gnome-shell to ask the user with a modal dialog what kind of device
was plugged in and switch the ports according to the user's response.

This adds a dependency on ALSA.

https://bugzilla.gnome.org/show_bug.cgi?id=755062