mouse, common: Fix Wacom buttons not receiving events
When configuring mice and touchpads, we might stumble upon a Wacom pad
device. To check whether it's a synaptics device, we open the device,
get the properties, and close it back up.

But when we close it, we end up releasing whatever grab the wacom plugin
might have had on the pad buttons:

           When a client makes an XCloseDevice request, any active grabs
           that the client has on the device are released.  Any event
           selections that the client has are deleted, as well as any
           passive grabs. If the requesting client is the last client
           accessing the device, the server may disable all access by X to
           the device.

As we cannot change that behaviour, we'll use a call that doesn't
require opening the device so we don't have to close it.

See https://bugs.freedesktop.org/show_bug.cgi?id=94487

https://bugzilla.gnome.org/show_bug.cgi?id=763092